a8n: Use request's actor when running CampaignPlan async
This fixes #6620 by taking the (if it's there) actor from the current
requests and passing it along with the `context.Background()` we use
when running a `CampaignPlan` in the background.

Both functions, `actor.FromContext` and `actor.WithActor`, are
`nil`-aware, so we're safe using them like and trying to pass on the
actor every time.