a8n: Ignore jobs with empty diff as completed in CampaignPlan status
I ran into this while testing.

When 2 CampaignJobs are created for two repositories and one CampaignJob
finishes successfully but has an empty diff it was counted towards the
"completed" count in the `campaignPlan.status`.

From an API consumer perspective, this can be confusing since it means
that this might be true:

    campaignPlan.changesets.totalCount != campaignPlan.status.completedCount

Since we already filter out changesets that have finished but have no
diff, with this change the following will always be true:

    campaignPlan.changesets.totalCount == campaignPlan.status.completedCount