Handle interval elapsing while Sync is running
Previously if this order of events happened we wouldn't instantly restart the
sync:

1. interval elapses
2. signal fires
3. Sync completes

This is because the signal firing isn't read since the goroutine monitoring it
stops after the interval elapses. So the sync will continue to run. Instead we
return to the older behaviour of sleeping after the sync is complete. To
acheive the behaviour we want we just make the sleep context aware.