a8n: Merge `a8n` and `run` packages (#6853)
* a8n: Merge `a8n` and `run` packages

* dev: Add enterprise/repo-updater to go-dbconn-import whitelist

After merging the `run` and `a8n` packages in 9ab8f4f6 this check
failed.

It's likely broken, since we only import `dbconn` in the test files in
the `a8n` package. But since enterprise/repo-updater _does_ access the
DB directly we can whitelist it here until we've fixed this check.