a8n: Consistent naming for "ref" and "ref name" (#6982)
* a8n: Rename `BaseRefName` to `BaseRef`

This is to be consistent, since git calls a ref that includes the
`refs/` prefix a ref, not a "ref name". A "ref name" often refers to the
abbreviate form, where instead of "refs/heads/master" it's "master".

* a8n: Be consistent in the naming of "ref"/"ref name"

If it includes the "refs/heads/" prefix, it's a "ref". If it doesn't,
it's the abbreviated version of the ref, sometimes called the "ref
name".