ci: Remove init-pipeline.yaml (#7013)
We now directly specify the generation step in the pipeline settings. This
removes a step from the critical path which had varying runtimes. For example
in the last buildkite build I looked at it took 21s and waited 2s.

Additionally we introduce a README lightly explaining how this script fits in.