search: correctly report zoekt timeouts through searchResultsCommon
`searchFilesInRepos` has a contract that timeouts are reported via the
`searchResultsCommon` structure. The unindexed searcher code in this
same function respects this, but the Zoekt code for some reason omitted
this.

Now this function appropriately returns timeouts to callers in the way
they would expect (e.g. even the GraphQL API and search pagination API
expect timeouts are reported in this way and not as a high-level error).