gitserver: Handle and ignore filepath.Walk error
filepath.Walk can return errors if we run into permission errors or a file
disappears between `readdir` and the `stat` of the file. In either case this
error isn't important for our cleanup, so we just ignore it. This is what we
typically do in the rest of the gitserver uses of filepath.Walk

Previously we ignored the error completely, which would lead to `info` being
nil => panic. This bug has existed for a long time, but is a hard to trip
panic.