repo-updater: Do not run source on deleted external services
If a user has deleted a service we don't need to check if the config can still
source the external service. For example a user may delete an external service
because the config is bad / the upstream code host is not reachable. In both
those cases the trigger call shouldn't fail.