repo-updater: Do not try to source deleted external services (#7035)
We now do not block deletion on triggering repo-updater syncs. We don't need
to give the user any feedback from repo-updater, so no need to wait for the
trigger to happen.

We do not try source deleted external services. We source for validation
purposes, but no point validating a deleted external service. We had bug
reports where we would present a warning that we failed to sync when deleting
an external service. However, this was expected since the external service
upstream code host was down, which is why the service was deleted.