Present search timeouts in a reasonable and consistent way; bump default Zoekt 1.5s timeout -> 3s (#7020)
* search: simplify context error check

* search: correctly report zoekt timeouts through searchResultsCommon

`searchFilesInRepos` has a contract that timeouts are reported via the
`searchResultsCommon` structure. The unindexed searcher code in this
same function respects this, but the Zoekt code for some reason omitted
this.

Now this function appropriately returns timeouts to callers in the way
they would expect (e.g. even the GraphQL API and search pagination API
expect timeouts are reported in this way and not as a high-level error).

* search: report search timeout alerts any time all repositories have timed out

Prior to this change the search timeout alert GUI was only relying on context
errors to present itself. We now present it anytime all repositories have timed
out (but not on partial timeouts, as that would prevent the frontend from
displaying partial results).

* search: improve timeout alert wording

* web: search: make search timeout recommendations consistent

Previously these recommendations:

- Were not displayed in two cases where a user would recieve a "please try a larger timeout:" message.
- Had a different, odd, yellow alert style in stark contrast to all other search timeout alerts.

* search: default zoekt timeout 1.5s -> 3s

* Update cmd/frontend/graphqlbackend/zoekt.go

* clarify logic + add documentation

* remove useless import

* eq eq eq

* fix test

* Update cmd/frontend/graphqlbackend/search_results.go

Co-Authored-By: Christina Forney <christina@sourcegraph.com>

* correct some poorly written tests

* do not pass the entire repository list to the client, just the size of it

* CHANGELOG: add 3.10.2 entries

* clarify wording