minversion: Lower to any go1.13 (#7083)
minversion is for ensuring developers are on a version of go which doesn't
introduce problems locally. It doesn't need to match the version we use in
CI. Currently minversion is higher than what is available from brew. So this
PR just lowers it to the actual minversion required for development purposes.