search: fix critical regression which broke diff, commit, and repo search (#7103)
In #7020 I failed to properly test that my change did not affect other
search types sufficiently. I tested text and symbol search, but failed to
manually test that commit/diff/repo search were not negatively affected.

Our end-to-end tests, regeression tests, and release process failed here
which I have filed at #7102 for us to address so this does not happen
again.

The cause of this issue was that diff/commit/repo search all were incorrectly
implemented in such a way that they did not properly fill out in `searchResultsCommon`
exactly which repositories they searched. Before my change, this would've
been a bug in our GraphQL API but after my change we were relying on
this to identify when timeouts ocurred.

This PR fixes the true root cause by correcting the search functions so
that they properly indicate which repositories they searched in accordance
with the searchResultsCommon and GraphQL API spec.

A patch release 3.10.3 with this change will be released ASAP.

Fixes #7060