support hiding global navbar's search input when useful
For pages that have their own saerch input that would conflict with or be confusing with the global search input, the GlobalNavbar component now accepts a `hideGlobalSearchInput` prop that hides the global search input. This is used in a subsequent commit for the new language statistics page.

Also makes other related changes to support testing GlobalNavbar.