a8n: Make campaign branch names globally unique
This commit makes campaign branch names globally unique and human
friendly at the same time. We convert the campaign name to an ASCII slug
and append the created at timestamp in unix seconds.

Using randomness in the name would prevent the retry mechanism to work
since we'd end up with a different headRefName in each attempt.

Fixes #7088