Revert to Fprint due to inconsistent lint check
The `Fprintln("...\n")` triggers https://github.com/golang/go/issues/18085#issuecomment-350383031, so using `Fprint("...\n")` instead. I'm changing it for both `Fprintf` calls in this block, even though the one can be `Fprintln` and the other `Fprint`, so that it's consistent and we don't see a weird mix of `Fprintln` and `Fprintf` here.