use a different redis cache for enhanced lang detection
This is necessary because when the server is using USE_ENHANCED_LANGUAGE_DETECTION, the stored results are different from when it's not. When that flag is disabled, all cached results will have line counts of 0. If the flag is later enabled, it will reuse cached results with 0 line counts and the overall result will be partially incorrect.