support previewing existing campaign plans
Adds a new GraphQL API field `CampaignPlan.previewURL` that returns a URL to the CampaignDetails page (`/campaigns/new?plan=...`) that shows the diffs and changesets for an existing campaign plan.

This makes it so that API clients (such as the `src` CLI) can create campaign plans programmatically but still allow users to visually preview them before creating the campaign.