Fix GitLab file actions spacing issue
GitLab reported that the Sourcegraph native integration messes up the spacing in their .file-actions toolbar: https://gitlab.com/gitlab-org/gitlab/issues/118445#note_260782106

This is due to a CSS rule in the _GitHub_ styles packaged in the browser extension / native integration, applying `display: flex` to `.file-actions` elements, which unfortunately also matches elements on GitLab.

Since I did not find a reasonable way to unambigually match `.file-actions` on GitHub only, I resorted to adding a class to it so that we can match it in our GitHub stylesheets without bleeding CSS to GitLab.