more closely dogfood our recommended LSIF action
Do not continue on error. If an error occurs, we want to see it because it means our users likely are seeing it, too. Note that we can still merge because LSIF is not a required PR check (so this change will not prevent us from fixing our site if it's down and the only way to fix it is by merging a new commit).