remove old saved searches redirects
The /search/queries and /search/searches paths were used 3+ months ago for saved searches. Those currently redirect to the user's saved searches page. The redirects have been in place for long enough and are not needed anymore. The cost of keeping them in place is that they clutter up the code and make refactoring harder.