remove unused pages (routes and components) (#7361)
* remove unused /open page (open-in-editor)

This is very old and was from when we had an experimental VS Code-based editor desktop application. It has not been used in 2+ years.

* remove old saved searches redirects

The /search/queries and /search/searches paths were used 3+ months ago for saved searches. Those currently redirect to the user's saved searches page. The redirects have been in place for long enough and are not needed anymore. The cost of keeping them in place is that they clutter up the code and make refactoring harder.

* remove unused editor auth routes

These have not been used in 2+ years. They used to be used for authenticating an experimental VS Code-based desktop editor product.

* remove unused repo graph routes

These were used 1.5+ years ago for an experimental (always feature-flagged off by default) repository section that showed dependencies and dependents.

* remove unused welcome/start page routes and redirects

The welcome and start pages were removed 6+ months ago in favor of https://about.sourcegraph.com. These routes and redirects are no longer needed.

* remove unused homepage diagram

This appears to be a wireframe of a possible homepage design we could have used a while ago. It is not needed.

* remove unused {sourcegraph.com => about.sourcegraph.com}/PATH redirects

- docs: about.sourcegraph.com/docs is invalid, no inlinks
- enterprise: sourcegraph.com/enterprise was only used very briefly, no inlinks
- beta: sourcegraph.com/beta was only used very briefly, no inlinks
- server: sourcegraph.com/server was only used very briefly, no inlinks

* remove or unexport unused TypeScript exports (#7362)

- Removes unused exports that are not used anywhere
- Unexports bindings that were only used in the same module

I used ts-unused-exports to produce an initial list and manually fixed all of the locations that seemed worthwhile.