rename Search{FilterChips => Scopes}
This is a better name for the component because it is only concerned with showing search scopes, not all kinds of search filter chips. (Search scopes are one kind of filter that are defined in settings. There are other kinds of filters, such as dynamic filters, that are not called "scopes".)