simplify SearchScopes component (#7382)
* rename Search{FilterChips => Scopes}

This is a better name for the component because it is only concerned with showing search scopes, not all kinds of search filter chips. (Search scopes are one kind of filter that are defined in settings. There are other kinds of filters, such as dynamic filters, that are not called "scopes".)

* simplify SearchScopes component

- Convert to functional component (from class component)
- Remove unneeded ISearchScope interface
- Remove handling of repogroup:sample that is not needed anymore
- Add React component snapshot test
- Remove unused contextual scopes (these were never actually triggered because this component is only shown on the homepage, not on any repository or tree pages)