remove remnants of repogroup:sample (#7383)
Now that Sourcegraph.com indexes ~10k public repositories, we don't need or recommend `repogroup:sample` anymore. This commit removes remnants of it from code and docs to avoid confusion.

It is extremely unlikely that any users actually used `repogroup:sample` and are relying on it, so this change is safe to make.