QueryBuilder: show on its own page, link from home
The query builder took up a lot of space on the home screen and had an unsightly link (that looked cluttered especially when there were search scopes, quick links, and/or notices). Now that we are introducing split search modes, we have a better place to link to the query builder: from the search mode menu.

To make the query builder link less cluttered for now (with splitSearchModes off), the link is slipped in beside the search scopes.