QueryBuilder: show on its own page, link from home (#7385)
* simplify SearchPage ordering of scopes and query builder

Previously, the order of SearchScopes and QueryBuilder differed.

- If there were search scopes, they were shown first (right below the query input).
- If there were no search scopes, the QueryBuilder was shown first.

There does not seem to be a strong argument for this, and it adds needless complexity to the code.

* web: move QueryBuilder to its own search subdir

The web/src/search/input dir was getting cluttered.

* QueryBuilder: factor out expand-collapse state

The QueryBuilder component now is only concerned with the query builder form, not the wrapper component around it to expand/collapse it based on localStorage.

* QueryBuilder: show on its own page, link from home

The query builder took up a lot of space on the home screen and had an unsightly link (that looked cluttered especially when there were search scopes, quick links, and/or notices). Now that we are introducing split search modes, we have a better place to link to the query builder: from the search mode menu.

To make the query builder link less cluttered for now (with splitSearchModes off), the link is slipped in beside the search scopes.