web: consistent bootstrap table padding, use bootstrap vars instead of magic constants
Instead of specifying `$table-cell-padding: 0.625rem`, we want to use a standard spacer so that table cells are aligned with (eg) card headers and footers. This means using a value expression of `map-get($spacers, 2)` or similar. However, `$spacers` is not defined until `@import 'bootstrap/scss/variables';`, so we need to move these var overrides into the `global-styles/tables.scss` file.