web: faster SearchResultsList test (#7451)
Previously this test took ~33s to run. Now it takes ~3s (which is about as good as it'll get because it uses the DOM). The reason it took so long was that the `resultCount` values were needlessly high, and the algorithm appears to be superlinear. The component itself does not appear to have the same performance problem; it only occurs using the test mocks (eg MockVisibilitySensor), probably because after-scroll callbacks are debounced in the browser automatically.