migrations: Enforce that repo.external is always set in DB
This has been the case for nearly a year in Sourcegraph, but we still have the
partial index from the days when it wasn't always set. This migration is a
step forward to ensure we enforce this behaviour and will allow us to remove
code which handles the unset case.

Note: On Sourcegraph.com we still had rows which didn't have the columns set
due to not running the normal repo-updater sync. I manually investigated those
rows and none seem "important". For the external data to be missing it implies
no one has visited the repo for many months. This delete 500k rows from our
repos table (still 4mil rows remaining).