db: remove Repos.Upsert
The repos table is only mutated by repo-updater. All call sites for
Repos.Upsert are either in tests or unused REST API endpoints. For tests I
moved the upsert code into a _test.go file so it is still available for
tests. For the unused API I grepped our org for calls to it, and it is unused.