search: use reposet to reduce work in indexed repohasfile filtering
Currently we search all indexed repositories on zoekt when filtering with
repohasfile. However, we can reduce that work and only search within the
current reposet.

On large instances this should increase performance whenever we are dealing
with a subset of all repos. This is quite common if a user does not have
permission to access all repositories. Additionally when a query contains a
repo filter or multiple repohasfile filters this will reduce the work done.