search: reduce boilerplate for indexed repohasfile tests (#7499)
This commit greatly reduces the boilerplate for the indexed repohasfile
tests. This is motivated by finding it hard to understand and extend the
current tests. Additionally it will give us the power to reproduce a current
issue in the implementation. Most of the commits are pretty straight-forward,
but the first is the most interesting. Instead of using hardcoded responses,
we create a simple zoekt searcher which can evaluate the queries we expect
from repohasfile filtering.