search: aggregate RepoURLs from zoekt replicas
When horizontal search was released this was assumed to be an unused
field. However, the repohasfile filter uses it. Not aggregating leads to
repohasfile filtering to always getting an empty set. So when using
repohasfilter you would never get any repos. When using -repohasfilter you
would get all repos. This is a regression since 3.10.