search: use reposet to reduce work in indexed repohasfile filtering (#7500)
* search: use reposet to reduce work in indexed repohasfile filtering

Currently we search all indexed repositories on zoekt when filtering with
repohasfile. However, we can reduce that work and only search within the
current reposet.

On large instances this should increase performance whenever we are dealing
with a subset of all repos. This is quite common if a user does not have
permission to access all repositories. Additionally when a query contains a
repo filter or multiple repohasfile filters this will reduce the work done.

* search: use set intersection for include repohasfile

This commit adjusts the implementation to reduce the number of cases it has to
handle. Now that we search within the current reposet, we can make assumptions
about the set of repos that is returned. This leads to a smaller
implementation.