repo-updater: Soft-delete repos in fake store (#7506)
This switches the implementation of our fake store to do soft deletion. All
our tests continue to work, and this is work towards supporting soft deleting
repos in the syncer. When we add soft delete to the DB store, we will add
tests that will fail unless soft deletion is working.