a8n: Ensure that same diff produces same commit SHA
Since `StartedAt` gets reset whenever we retry the creation of a
Campaign (or when we update a Campaign later) we produce different
commit SHAs anytime this runs, since one of the "ingredients" changed.

And since #7512 ensures that we can run now push new commits to the same
ref, we want to avoid producing and pushing new commits with the same
diff to the codehost.