gitserver: Allow updating ref with new commits (#7512)
Before this change it wasn't possible to use 'createCommitFromPatch'
twice with the same 'TargetRef' and different patches. When trying to
push the new commit to the codehost (tested with GitHub) it would report
that everything's up to date.

The reason was that we created ambigious refs. Example: with a
'TargetRef' of 'A' the call to 'git update-ref -- A <commitsha>' created
a ref called 'A'. But when pushing to the codehost with 'git push
A:refs/heads/A' we also implicitly created a second ref, called
'refs/heads/A', making 'A' as a ref name ambigious.

When the endpoint was called _again_ with a different diff a new commit
was created and the 'A' ref was updated, but since the name was
ambigious the 'refs/heads/A' ref was used, which was correctly reported
as being up to date with the remote host.

This could be confirmed with 'git show-ref' and 'git show' which showed
different commit shaws and a "refname is ambiguous" warning.

What this change here does is

1. avoid ambigous ref name by ensuring that the ref has a 'refs/heads/'
   prefix when 'Push: true'
2. pushing 'cmtHash' (instead of the 'ref') to the full ref name on the
   codehost _before_ updating the local ref
3. updating the local full ref name to the now-pushed commit

That allows us to now use 'createCommitFromPatch' with the same
'TargetRef' to push changed diffs to the codehost.