auth: GitHub authentication fixes
This PR fixes a regression introduced in #7331 and reported in #7518 as
well setting a well-typed default for the new `allowOrgs` setting.

The issue was that `oauth.getProvider` does a concrete type assertion
and in the previous PR we changed what was stored in as the interface
implementation to a wrapper type.

Fixes #7518