enterprise/a8n: Move processing campaign jobs to a background routine
Instead of launching goroutines for each campaign we now persist the
jobs to the DB first and a background process will pick them up. This
protects us against jobs getting lost if we lose a frontend instance.

The implementation uses row level locking in PostgreSQL to ensure that
only one routine per instance can work on a job at any time. The
downside is that it requires polling the database for new jobs rather
than working on them immediatley.