repo-updater: db store always inserts verbatim repo fields and only returns id on insert (#7612)
* remove coalesce in repo update statement

external ID's are always set, so the coalesce is essentially a noop. The
coalesce for created_at and updated_at is also a noop, since it is always set.

* remove external id nullifs

They are always set

* do not return rows in update or delete statement

We don't need to return, we already know the ID.

* only scan ID on insert

We don't need to scan every field, since we directly insert it.