httpcli: introduce ExternalTransportOpt and site setting
We introduce the site setting tls.external which administrators can configure
for all external communication. We also add ExternalTransportOpt which allows
an httpcli.Factory to be configured to update the transport with the settings
specified in tls.external.

This commit just introduces the "insecureSkipVerify". This is to minimize the
change. Future commits will introduce options to configure
RootCAs/etc. Additionally ExternalTransportOpt is not used yet, but will be in
future commits.