More campaign UI refactors and minor changes (#7447)
* a8n: use new ImageAutoAdjustIcon icon for campaigns

This icon better conveys "fix" and "auto" than the current icon (which is of a burndown chart going down).

Also upgrades mdi-react and fixes snapshots to not depend on the particular SVG definition (which can fluctuate meaninglessly when mdi-react is updated).

* a8n: remove namespace field, always use current user

Right now, the namespace field in the campaign form is not useful. Site admins always see all campaigns on the campaign list page, and non-site admins can't view any campaigns at all. It adds needless complexity to the form. This commit makes all campaigns created in the current user's namespace. When we support campaign namespaces throughout the UI, we can add this field back (or something like it).

* a8n: do not show username in new campaign form

It is not useful.

* a8n: improve placeholders for Campaign{Title,Description}Field components

* a8n: encapsulate campaign plan spec form data in component

* a8n: use auto-resizing textarea for campaign description

* a8n: merge rendering logic for Campaign{,Plan} changesets

* a8n: merge rendering logic for Campaign{,Plan} diffs

* a8n: extract CampaignTabs