search: support searching text in multiple refs specified by glob
Adds support for searching for text in multiple refs per repository where the refs are matched by glob. Previously you needed to explicitly list each revspec you wanted to search.

To use this, give a query of the form `repo:myrepo@*refs/heads/:^refs/heads/myexclude`. That will search all branches except for the `myexclude` branch. The glob syntax is the same as used for diff search (which, in turn, is the same as used by `git log --glob ... --exclude ...`).

This is all behind the searchMultipleRevisionsPerRepository site config feature flag (disabled by default).