a8n: When retrying Campaign creation only retry failed jobs (#7615)
This was an oversight on my part. I didn't check the opposite case in
the test.

The bug manifested itself in that we retried every job when retrying
Campaign creation, even the ones that successfully created a changeset
on a codehost. That wasn't a huge problem since that function is
idempotent, but it's still needless work and the behavior of the method
doesn't match its name.