a8n: Persistent queues and workers for CampaignJob (#7614)
* enterprise/a8n: Move processing campaign jobs to a background routine

Instead of launching goroutines for each campaign we now persist the
jobs to the DB first and a background process will pick them up. This
protects us against jobs getting lost if we lose a frontend instance.

The implementation uses row level locking in PostgreSQL to ensure that
only one routine per instance can work on a job at any time. The
downside is that it requires polling the database for new jobs rather
than working on them immediatley.

* enterprise/a8n: jobTimeout was already specified and used in runJob

* enterprise/a8n: Extract campaign type strings into constants

* enterprise/a8n: Simplify runner tests.

Don't try and run background process. A separate test will test this
functionality.

* enterprise/a8n: Add test for ConsumePendingCampaignJobs

* enterprise/a8n: Add a test for ProcessPendingCampaignJob

* enterprise/a8n: Add comment for change in runner.Wait()

* enterprise/a8n: Sort pending jobs by id

It'll be faster than by date and will result in the same ordering

* enterprise/a8n: Rename function

* enterprise/a8n: Rename runner.Run to runner.Persist

And update comments

* enterprise/a8n: Update comment

* enterprise/a8n: Simplify background worker

* enterprise/a8n: Reimplement runner.Wait()

It now polls the DB for finished jobs

* enterprise/a8n: Update comment

* enterprise/a8n: Use store.GetCampaignPlanStatus

* enterprise/a8n: Move background process

From enterprise/frontend to enterprise/repo-updater

* Revert "enterprise/a8n: Move background process"

This reverts commit f0a86df5fa783161f6977cec3b0ac96bce28622f.

* enterprise/a8n: Remove unused variable

* enterprise/a8n: Reduce nesting

* enterprise/a8n: Use context in RunCampaignJobs

For cancellation instead of doneChan

* enterprise/a8n: Improve comment

* enterprise/a8n: Rename runner.Persist to CreatePlanAndJobs

* enterprise/a8n: Fix test race

* enterprise/a8n: Fix lint warning